Add commandline arg to output inlining to stdout instead of to file

+++ Program.cs
@@ -12,7 +12,7 @@ if (!result.Errors.Any() && args.Length > 0)
 
     if (Options.IncludeSections?.Any() ?? false)
         sectionName = Options.IncludeSections.First();
-    
+
 
     var ParseResult = Parse(File.ReadAllText(Options.Filename));
 
@@ -30,26 +30,34 @@ if (!result.Errors.Any() && args.Length > 0)
             var ShortName = Path.GetFileNameWithoutExtension(Options.Filename);
             var LongFilename = Options.Filename[..^Extension.Length];
 
-            Console.Out.WriteLine($"Assembled {ShortName} => {ShortName}.{sectionName}{Extension}");
-            Console.Out.WriteLine($"{AssemblyResult.FinalSections.Count} sections totalling {AssemblyResult.OutputLines.Count} line{(AssemblyResult.OutputLines.Count != 1 ? "s" : "")} of code");
-            File.WriteAllText($"{LongFilename}.{sectionName}{Extension}", AssemblyResult.Output);
-
-            if (Options.EmitSymbolFile)
+            if (!Options.OutputToSTDOUT)
             {
-                File.Delete($"{LongFilename}.{sectionName}.sym");
-                using var SymbolFile = File.OpenWrite($"{LongFilename}.{sectionName}.sym");
-                using StreamWriter writer = new(SymbolFile);
-                var last_section = "";
-                foreach (var Symbol in AssemblyResult.Symbols.Where(Symbol => AssemblyResult.FinalSections.Contains(Symbol.Section)))
-                {
-                    if (last_section != Symbol.Section.Name)
-                        writer.WriteLine($"section {Symbol.Section.Name} offset {Symbol.Section.Offset}");
+                Console.Out.WriteLine($"Assembled {ShortName} => {ShortName}.{sectionName}{Extension}");
+                Console.Out.WriteLine(
+                    $"{AssemblyResult.FinalSections.Count} sections totalling {AssemblyResult.OutputLines.Count} line{(AssemblyResult.OutputLines.Count != 1 ? "s" : "")} of code");
+                File.WriteAllText($"{LongFilename}.{sectionName}{Extension}", AssemblyResult.Output);
 
-                    last_section = Symbol.Section.Name;
-                    writer.WriteLine($"  {Symbol.SymbolName} offset {Symbol.Value}");
+
+                if (Options.EmitSymbolFile)
+                {
+                    File.Delete($"{LongFilename}.{sectionName}.sym");
+                    using var SymbolFile = File.OpenWrite($"{LongFilename}.{sectionName}.sym");
+                    using StreamWriter writer = new(SymbolFile);
+                    var last_section = "";
+                    foreach (var Symbol in AssemblyResult.Symbols.Where(Symbol => AssemblyResult.FinalSections.Contains(Symbol.Section)))
+                    {
+                        if (last_section != Symbol.Section.Name)
+                            writer.WriteLine($"section {Symbol.Section.Name} offset {Symbol.Section.Offset}");
+
+                        last_section = Symbol.Section.Name;
+                        writer.WriteLine($"  {Symbol.SymbolName} offset {Symbol.Value}");
+                    }
                 }
             }
-
+            else
+            {
+                Console.Out.Write(AssemblyResult.Output);
+            }
         }
         else
         {
@@ -70,9 +78,11 @@ if (!result.Errors.Any() && args.Length > 0)
         foreach (var error in ParseResult.Errors)
             Console.Error.WriteLine($"Error: {error}");
     }
+
     foreach (var warning in ParseResult.Warnings)
         Console.Error.WriteLine($"Warning: {warning}");
 
     return Failed ? 1 : 0;
 }
+
 return -1;
